workshops/04_workshop - Agent to Docker/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
ddgs==9.9.2
dnspython==2.8.0
duckduckgo_search==8.1.1
email-validator==2.3.0
fake-useragent==2.2.0
fastapi==0.123.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
Hypercorn==0.18.0
hyperframe==6.1.0
idna==3.11
lxml==6.0.2
primp==0.15.0
priority==2.0.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.2.5
requests==2.32.5
rsa==4.9.1
socksio==1.0.0
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
wsproto==1.3.2